Move remote_update to common and use in both iso and vmx
