+++ builder/vmware/iso/builder.go
@@ -88,14 +88,18 @@ func (b *Builder) Run(ctx context.Context, ui packer.Ui, hook packer.Hook) (pack
 			Directories: b.config.FloppyConfig.FloppyDirectories,
 			Label:       b.config.FloppyConfig.FloppyLabel,
 		},
-		&stepRemoteUpload{
+		&vmwcommon.StepRemoteUpload{
 			Key:       "floppy_path",
 			Message:   "Uploading Floppy to remote machine...",
 			DoCleanup: true,
+			Checksum:     "",
+			ChecksumType: "none",
 		},
-		&stepRemoteUpload{
+		&vmwcommon.StepRemoteUpload{
 			Key:     "iso_path",
 			Message: "Uploading ISO to remote machine...",
+			Checksum:     b.config.ISOChecksum,
+			ChecksumType: b.config.ISOChecksumType,
 		},
 		&stepCreateDisk{},
 		&stepCreateVMX{},
